requirements.txt
# main packages
django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7

# third party packages
djangorestframework_simplejwt
markdown
django-filter
django-decouple
django-cors-headers
pillow

# linting and test packages
flake8
black

# api docs
drf-yasg
drf-yasg[validation]